DOMA-1259 split logic of the TicketCard into two components
